refactor(core): update minimap slider controller design
affects: @medly-components/core, @medly-components/theme
