--- packages/core/src/components/Table/Minimap/SliderController/index.ts
@@ -1,3 +1,0 @@
-import { SliderController } from './SliderController';
-
-export default SliderController;
