--- packages/core/src/components/Table/Minimap/SliderController/types.ts
@@ -1,3 +1,0 @@
-export interface Props {
-    controllerWidth: number;
-}
